Moved `Version` to the net.dhleong.acl.util package
