+++ src/net/dhleong/acl/iface/ThreadedArtemisNetworkInterface.java
@@ -15,10 +15,10 @@ import net.dhleong.acl.enums.ConnectionType;
 import net.dhleong.acl.protocol.ArtemisPacket;
 import net.dhleong.acl.protocol.ArtemisPacketException;
 import net.dhleong.acl.protocol.Protocol;
-import net.dhleong.acl.protocol.Version;
 import net.dhleong.acl.protocol.core.setup.VersionPacket;
 import net.dhleong.acl.protocol.core.setup.WelcomePacket;
 import net.dhleong.acl.protocol.core.world.ObjectUpdatePacket;
+import net.dhleong.acl.util.Version;
 import net.dhleong.acl.world.ArtemisObject;
 
 /**
